Initial implementation
This is cribbed somewhat from Snarfetch.  We have our own directory
traversal code because `ignore` quite sensibly doesn't want anything
to do with the filesystem and other libraries I found don't handle
.gitignore files properly if they're not in the project root.

Maybe I'll pull that code into a separate library if anyone wants it,
for now it does what I need it too and I don't need it anywhere else.

This code could probably do with a refactor, but it "works" so I want to
commit it.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>